script/requirements.txt
google-cloud
requests